# Preserve the order of right table in NestedLoopJoinExec