## Which issue does this PR close?

Closes #11332.

## Rationale for this change

See the linked issue.

## What changes are included in this PR?

`NestedLoopJoinExec` now preserves the order of the right table for `Inner`, `Right`, `RightAnti` and `RightSemi` join types.

## Are these changes tested?

Tested with `join_maintains_right_order` and relevant SQL logic tests

## Are there any user-facing changes?

This PR changes the output ordering of some queries with `JOIN`, as reflected in SQL logic tests. The output values and the API remain unchanged.