@comphead 2024-09-18
Thanks @korowa for checking this, the join time is 9x now @berkaysynnada @alihan-synnada