@comphead 2024-09-18
Filed https://github.com/apache/datafusion/issues/12528