@korowa 2024-09-18
If this PR enforces ordering by the price of slowdown for all types of inputs (the example shows fairly small build side input, but it also may be enough), it, perhaps, should be reverted and reimplemented -- I suppose, there exist more compromise solutions for this issue, since, if the root cause of slowdown demonstrated in the example, is increased number calls to `apply_join_filter_to_indices`, the problem, perhaps, could be solved by something like #8952 (which could accumulate indices up to batch size and only then apply filters).