@berkaysynnada 2024-09-19
Thanks for reporting this @korowa. I believe it’s possible to maintain the order without affecting performance. We will work on fixing this shortly.